FIX: compiler warning
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@2674 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/apps/filter.cxx
@@ -228,7 +228,7 @@ doMain( const int argc, const char* argv[] )
         {
         if ( Coupe ) 
           {
-          cmtk::TypedArray::SmartPtr filtered( cmtk::FilterVolume::CoupeFilter( volume, CoupeWindowRadius, CoupeBeta ) );
+          cmtk::TypedArray::SmartPtr filtered( cmtk::FilterVolume::CoupeFilter( volume, static_cast<int>( CoupeWindowRadius ), CoupeBeta ) );
           volume->SetData( filtered );
           }
         }
